pyproject.toml
[tool.poetry]
name = "sslex"
version = "0.1.0"
description = "SSL Expiry checker and reminder tool"
authors = ["Jagadeeswaran T"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
python-whois = "^0.8.0"
requests = "^2.28.1"
pyOpenSSL = "^22.0.0"
fake-useragent = "^0.1.11"
huey = "^2.4.4"
redis = "^4.3.4"
mypy = {git = "https://github.com/TheGardenMan/mypy.git"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
black = {version = "^22.6.0", allow-prereleases = true}
ipython = "^8.4.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
